Repeated type annotations in docstrings removed; data store and data opener split in two seperated files
